refactor: Rename SectionEditModal in SectionEditDialog
to be consistent with TripDialog
