+++ src/components/SectionEditDialog/ModeList.jsx
@@ -2,7 +2,7 @@ import React from 'react'
 
 import List from 'cozy-ui/transpiled/react/MuiCozyTheme/List'
 
-import ModeItem from 'src/components/SectionEditModal/ModeItem'
+import ModeItem from 'src/components/SectionEditDialog/ModeItem'
 import { modes } from 'src/components/helpers'
 
 const ModeList = ({ section, closeModal }) => {
