+++ src/components/SectionEditDialog/helpers.js
