+++ src/components/SectionEditDialog/index.js
@@ -1,0 +1,3 @@
+import SectionEditDialog from './SectionEditDialog'
+
+export default SectionEditDialog
