--- src/components/SectionEditModal/index.js
@@ -1,3 +1,0 @@
-import SectionEditModal from './SectionEditModal'
-
-export default SectionEditModal
